optimized the code
