+++ src/main/java/com/google/codeu/servlets/AboutMeServlet.java
@@ -1,8 +1,6 @@
 package com.google.codeu.servlets;
 
 import java.io.IOException;
-import java.util.List;
-import java.util.Map;
 
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -11,10 +9,6 @@ import javax.servlet.http.HttpServletResponse;
 
 
 import com.github.rjeschke.txtmark.Processor;
-import com.google.appengine.api.blobstore.*;
-import com.google.appengine.api.images.ImagesService;
-import com.google.appengine.api.images.ImagesServiceFactory;
-import com.google.appengine.api.images.ServingUrlOptions;
 import com.google.appengine.api.users.UserService;
 import com.google.appengine.api.users.UserServiceFactory;
 import com.google.codeu.data.Datastore;
@@ -24,6 +18,8 @@ import org.jsoup.Jsoup;
 import org.jsoup.safety.Whitelist;
 import org.kefirsf.bb.BBProcessorFactory;
 import org.kefirsf.bb.TextProcessor;
+
+import static com.google.codeu.servlets.BlobstoreUploadUrlServlet.getUploadedFileUrl;
 
 /**
  * Handles fetching and saving user data.
@@ -59,7 +55,6 @@ public class AboutMeServlet extends HttpServlet {
         //Load the user details and send it as the json
         Gson gson = new Gson();
         String json = gson.toJson(userData);
-        System.out.println(json);
         response.getOutputStream().println(json);
     }
 
@@ -72,8 +67,6 @@ public class AboutMeServlet extends HttpServlet {
             response.sendRedirect("/index.html");
             return;
         }
-
-
 
         String userEmail = userService.getCurrentUser().getEmail();
         String aboutMe = Jsoup.clean(request.getParameter("aboutMe"), Whitelist.none());
@@ -88,48 +81,16 @@ public class AboutMeServlet extends HttpServlet {
 
         // Get the URL of the image that the user uploaded to Blobstore.
         String imageUrl = getUploadedFileUrl(request, "image");
-
+        User user = new User(userEmail, aboutMe);
+        user.setNickName(nickName);
         if (imageUrl != null) {
-            User user = new User(userEmail, aboutMe, nickName, imageUrl);
+           user.setImageUrl(imageUrl);
             datastore.storeUser(user);
         } else {
-            User user = new User(userEmail, aboutMe, nickName);
             datastore.storeUser(user);
         }
 
         response.sendRedirect("/user-page.html?user=" + userEmail);
     }
 
-    /**
-     * Returns a URL that points to the uploaded file, or null if the user didn't upload a file.
-     */
-    private String getUploadedFileUrl(HttpServletRequest request, String formInputElementName) {
-        BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
-        Map<String, List<BlobKey>> blobs = blobstoreService.getUploads(request);
-        List<BlobKey> blobKeys = blobs.get("image");
-
-        // User submitted form without selecting a file, so we can't get a URL. (devserver)
-        if (blobKeys == null || blobKeys.isEmpty()) {
-            return null;
-        }
-
-        // Our form only contains a single file input, so get the first index.
-        BlobKey blobKey = blobKeys.get(0);
-
-        // User submitted form without selecting a file, so we can't get a URL. (live server)
-        BlobInfo blobInfo = new BlobInfoFactory().loadBlobInfo(blobKey);
-        if (blobInfo.getSize() == 0) {
-            blobstoreService.delete(blobKey);
-            return null;
-        }
-
-        // We could check the validity of the file here, e.g. to make sure it's an image file
-        // https://stackoverflow.com/q/10779564/873165
-
-        // Use ImagesService to get a URL that points to the uploaded file.
-        ImagesService imagesService = ImagesServiceFactory.getImagesService();
-        ServingUrlOptions options = ServingUrlOptions.Builder.withBlobKey(blobKey);
-        return imagesService.getServingUrl(options);
-    }
-    
 }
