+++ src/main/java/com/google/codeu/servlets/MessageServlet.java
@@ -17,10 +17,6 @@
 package com.google.codeu.servlets;
 
 import com.github.rjeschke.txtmark.Processor;
-import com.google.appengine.api.blobstore.*;
-import com.google.appengine.api.images.ImagesService;
-import com.google.appengine.api.images.ImagesServiceFactory;
-import com.google.appengine.api.images.ServingUrlOptions;
 import com.google.appengine.api.users.UserService;
 import com.google.appengine.api.users.UserServiceFactory;
 import com.google.codeu.data.Datastore;
@@ -32,7 +28,6 @@ import com.google.cloud.language.v1.Sentiment;  //sentiment
 
 import java.io.IOException;
 import java.util.List;
-import java.util.Map;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -43,6 +38,8 @@ import org.jsoup.Jsoup;
 import org.jsoup.safety.Whitelist;
 import org.kefirsf.bb.BBProcessorFactory;
 import org.kefirsf.bb.TextProcessor;
+
+import static com.google.codeu.servlets.BlobstoreUploadUrlServlet.getUploadedFileUrl;
 
 /**
  * Handles fetching and saving {@link Message} instances.
@@ -118,49 +115,18 @@ public class MessageServlet extends HttpServlet {
 
         // Get the URL of the image that the user uploaded to Blobstore.
         String imageUrl = getUploadedFileUrl(request, "image");
-        System.out.println("imageUrl "+imageUrl);
+        Message message = new Message(user,text,score);
         if(imageUrl!=null) {
-            Message message = new Message(user, text,imageUrl,score);
+            message.setImageUrl(imageUrl);
             datastore.storeMessage(message);
         }
         else {
-            Message message = new Message(user,text,score);
             datastore.storeMessage(message);
         }
         response.sendRedirect("/user-page.html?user=" + user);
     }
 
 
-    /**
-     * Returns a URL that points to the uploaded file, or null if the user didn't upload a file.
-     */
-    private String getUploadedFileUrl(HttpServletRequest request, String formInputElementName) {
-        BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
-        Map<String, List<BlobKey>> blobs = blobstoreService.getUploads(request);
-        List<BlobKey> blobKeys = blobs.get("image");
 
-        // User submitted form without selecting a file, so we can't get a URL. (devserver)
-        if (blobKeys == null || blobKeys.isEmpty()) {
-            return null;
-        }
-
-        // Our form only contains a single file input, so get the first index.
-        BlobKey blobKey = blobKeys.get(0);
-
-        // User submitted form without selecting a file, so we can't get a URL. (live server)
-        BlobInfo blobInfo = new BlobInfoFactory().loadBlobInfo(blobKey);
-        if (blobInfo.getSize() == 0) {
-            blobstoreService.delete(blobKey);
-            return null;
-        }
-
-        // We could check the validity of the file here, e.g. to make sure it's an image file
-        // https://stackoverflow.com/q/10779564/873165
-
-        // Use ImagesService to get a URL that points to the uploaded file.
-        ImagesService imagesService = ImagesServiceFactory.getImagesService();
-        ServingUrlOptions options = ServingUrlOptions.Builder.withBlobKey(blobKey);
-        return imagesService.getServingUrl(options);
-    }
 }
