+++ src/main/webapp/feed.html
@@ -21,6 +21,7 @@
     <script src="/js/navigation-loader.js"></script>
 
     <!-- Custom scripts -->
+    <script src="/js/message-tamplate.js"></script>
     <script src="/js/feed-page-loader.js"></script>
 </head>
