+++ src/main/webapp/user-page.html
@@ -39,6 +39,7 @@ limitations under the License.
     <script src="/js/navigation-loader.js"></script>
 
     <!-- Custom scripts-->
+    <script src="/js/message-tamplate.js"></script>
     <script src="/js/user-page-loader.js"></script>
